requirements.txt
asgiref==3.8.1
Babel==2.15.0
click==8.1.7
colorama==0.4.6
Django==5.0.6
djangorestframework==3.15.1
json5==0.9.25
sqlparse==0.5.0
tqdm==4.66.4
tzdata==2024.1
webencodings==0.5.1
